+++ ThirdPartyAdapters/liftoffmonetize/liftoffmonetize/src/test/java/com/google/ads/mediation/vungle/waterfall/VungleWaterfallBannerAdTest.kt
@@ -1,11 +1,24 @@
-package com.google.ads.mediation.vungle.rtb
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.ads.mediation.vungle.waterfall
 
 import android.content.Context
 import androidx.core.os.bundleOf
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.ads.mediation.adaptertestkit.AdErrorMatcher
-import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_APP_ID
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_PLACEMENT_ID
 import com.google.ads.mediation.adaptertestkit.createMediationBannerAdConfiguration
@@ -21,6 +34,7 @@ import com.vungle.ads.BaseAd
 import com.vungle.ads.VungleBannerView
 import com.vungle.ads.VungleError
 import com.vungle.ads.internal.protos.Sdk.SDKError
+import kotlin.use
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -33,12 +47,12 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 
-/** Tests for [VungleRtbBannerAd]. */
+/** Tests [VungleWaterfallBannerAd]. */
 @RunWith(AndroidJUnit4::class)
-class VungleRtbBannerAdTest {
+class VungleWaterfallBannerAdTest {
 
   /** Unit under test. */
-  private lateinit var adapterRtbBannerAd: VungleRtbBannerAd
+  private lateinit var adapterWaterfallBannerAd: VungleWaterfallBannerAd
 
   private val context = ApplicationProvider.getApplicationContext<Context>()
   private val bannerAdCallback = mock<MediationBannerAdCallback>()
@@ -51,7 +65,7 @@ class VungleRtbBannerAdTest {
   private val baseAd = mock<BaseAd>()
   private val vungleFactory =
     mock<VungleFactory> { on { createBannerAd(any(), any(), any()) } doReturn vungleBannerView }
-  private val mediationBannerAdConfiguration =
+  private val bannerAdConfig =
     createMediationBannerAdConfiguration(
       context = context,
       serverParameters =
@@ -59,12 +73,11 @@ class VungleRtbBannerAdTest {
           VungleConstants.KEY_APP_ID to TEST_APP_ID,
           VungleConstants.KEY_PLACEMENT_ID to TEST_PLACEMENT_ID,
         ),
-      bidResponse = AdapterTestKitConstants.TEST_BID_RESPONSE,
     )
 
   @Before
   fun setUp() {
-    adapterRtbBannerAd = VungleRtbBannerAd(bannerAdLoadCallback, vungleFactory)
+    adapterWaterfallBannerAd = VungleWaterfallBannerAd(bannerAdLoadCallback, vungleFactory)
 
     doAnswer { invocation ->
         val args: Array<Any> = invocation.arguments
@@ -78,12 +91,12 @@ class VungleRtbBannerAdTest {
   fun onAdLoaded_addsLiftoffBannerViewToBannerLayoutAndCallsLoadSuccess() {
     mockStatic(VungleInitializer::class.java).use {
       whenever(VungleInitializer.getInstance()) doReturn mockVungleInitializer
-      adapterRtbBannerAd.validateParamsAndLoadAd(mediationBannerAdConfiguration)
+      adapterWaterfallBannerAd.validateParamsAndLoadAd(bannerAdConfig)
     }
 
-    adapterRtbBannerAd.onAdLoaded(baseAd)
+    adapterWaterfallBannerAd.onAdLoaded(baseAd)
 
-    verify(bannerAdLoadCallback).onSuccess(adapterRtbBannerAd)
+    verify(bannerAdLoadCallback).onSuccess(adapterWaterfallBannerAd)
   }
 
   @Test
@@ -94,7 +107,7 @@ class VungleRtbBannerAdTest {
         on { errorMessage } doReturn "Liftoff Monetize SDK banner ad load failed."
       }
 
-    adapterRtbBannerAd.onAdFailedToLoad(baseAd, liftoffError)
+    adapterWaterfallBannerAd.onAdFailedToLoad(baseAd, liftoffError)
 
     val expectedError =
       AdError(liftoffError.code, liftoffError.errorMessage, VUNGLE_SDK_ERROR_DOMAIN)
@@ -104,16 +117,16 @@ class VungleRtbBannerAdTest {
   private fun renderAdAndMockLoadSuccess() {
     mockStatic(VungleInitializer::class.java).use {
       whenever(VungleInitializer.getInstance()) doReturn mockVungleInitializer
-      adapterRtbBannerAd.validateParamsAndLoadAd(mediationBannerAdConfiguration)
+      adapterWaterfallBannerAd.validateParamsAndLoadAd(bannerAdConfig)
     }
-    adapterRtbBannerAd.onAdLoaded(baseAd)
+    adapterWaterfallBannerAd.onAdLoaded(baseAd)
   }
 
   @Test
   fun onAdClicked_reportsAdClickedAndAdOpened() {
     renderAdAndMockLoadSuccess()
 
-    adapterRtbBannerAd.onAdClicked(baseAd)
+    adapterWaterfallBannerAd.onAdClicked(baseAd)
 
     verify(bannerAdCallback).reportAdClicked()
     verify(bannerAdCallback).onAdOpened()
@@ -123,7 +136,7 @@ class VungleRtbBannerAdTest {
   fun onAdImpression_reportsAdImpression() {
     renderAdAndMockLoadSuccess()
 
-    adapterRtbBannerAd.onAdImpression(baseAd)
+    adapterWaterfallBannerAd.onAdImpression(baseAd)
 
     verify(bannerAdCallback).reportAdImpression()
   }
@@ -132,21 +145,21 @@ class VungleRtbBannerAdTest {
   fun onAdLeftApplication_callsOnAdLeftApplication() {
     renderAdAndMockLoadSuccess()
 
-    adapterRtbBannerAd.onAdLeftApplication(baseAd)
+    adapterWaterfallBannerAd.onAdLeftApplication(baseAd)
 
     verify(bannerAdCallback).onAdLeftApplication()
   }
 
   @Test
   fun onAdEnd_noCrash() {
-    adapterRtbBannerAd.onAdEnd(baseAd)
+    adapterWaterfallBannerAd.onAdEnd(baseAd)
 
     // No matching callback exists on the GMA SDK. This test just verifies that there was no crash.
   }
 
   @Test
   fun onAdStart_noCrash() {
-    adapterRtbBannerAd.onAdStart(baseAd)
+    adapterWaterfallBannerAd.onAdStart(baseAd)
 
     // No matching callback exists on the GMA SDK. This test just verifies that there was no crash.
   }
@@ -159,7 +172,7 @@ class VungleRtbBannerAdTest {
         on { errorMessage } doReturn "Liftoff Monetize SDK banner ad play failed."
       }
 
-    adapterRtbBannerAd.onAdFailedToPlay(baseAd, liftoffError)
+    adapterWaterfallBannerAd.onAdFailedToPlay(baseAd, liftoffError)
 
     // No matching callback exists on the GMA SDK. This test just verifies that there was no crash.
   }